+++ src/converse-minimize.js
@@ -12,7 +12,7 @@ import tpl_toggle_chats from "templates/toggle_chats.html";
 import tpl_trimmed_chat from "templates/trimmed_chat.html";
 
 
-const { _ , Backbone, Promise, Strophe, moment } = converse.env;
+const { _ , Backbone, Promise, Strophe, dayjs } = converse.env;
 const u = converse.env.utils;
 
 converse.plugins.add('converse-minimize', {
@@ -51,7 +51,7 @@ converse.plugins.add('converse-minimize', {
                 }
                 this.save({
                     'minimized': this.get('minimized') || false,
-                    'time_minimized': this.get('time_minimized') || moment(),
+                    'time_minimized': this.get('time_minimized') || dayjs(),
                 });
             },